Make use of default primitive values set a strict type
Closes #38
