+++ test/test.js
@@ -2,3 +2,4 @@ require("./define-class-test");
 require("./define-test");
 require("./define-types-test");
 require("./type-events-test");
+require("./define-class-default-test");
